+++ src/components/flight/VcsFlightAnchorsComponent.vue
@@ -58,6 +58,7 @@
     </v-container>
   </VcsFormSection>
 </template>
+
 <script>
   import { computed, inject, onMounted, onUnmounted, ref } from 'vue';
   import { VContainer, VSheet, VSpacer } from 'vuetify/components';